Fix compilation when `TORCHDYNAMO_DEBUG` for Python 3.10.x and 3.11.x
The frame struct changes in each release of Python, so the debugging
logic that prints out properties of it needs to be adapted.

The code as is compiles well in Python 3.9.x and 3.8.x. For future
versions the following adaptations are needed:

Python 3.10.x:
Property `f_executing` is gone and `f_state` should be used. [PyFrameState](https://github.com/python/cpython/blob/3.10/Include/cpython/frameobject.h#L47)

Python 3.11.x:
The structure representing a frame changes a bit more here so some changes are needed:
- to get the frame state one needs to use `_PyFrame_GetGenerator(frame)->gi_frame_state`.
- `f_lineno` is now accessible via `frame->frame_obj->f_lineno.
- `f_lasti` can get accessed via `PyFrame_GetLasti(frame->frame_ojb).
- `f_iblock` is gone.
- More details in the [doc](https://github.com/python/cpython/blob/3.11/Doc/c-api/frame.rst).